PORTING THE MOD: PART 2

+++ src/main/java/com/aqutheseal/celestisynth/common/attack/aquaflora/AquafloraPetalPiercesAttack.java
@@ -53,9 +53,9 @@ public class AquafloraPetalPiercesAttack extends AquafloraAttack {
             getPlayer().playSound(CSSoundEvents.CS_AIR_SWING.get(), 0.25F, 1.3F + getPlayer().level().random.nextFloat());
             CSEffectEntity.createInstance(player, null, CSVisualTypes.AQUAFLORA_STAB.get(), -0.5 + getPlayer().level().random.nextDouble() + calculateXLook(player) * 3, (-0.5 + getPlayer().level().random.nextDouble()) + (2 + calculateYLook(player) * 3), -0.5 + getPlayer().level().random.nextDouble() + calculateZLook(player) * 3);
 
-            List<Entity> entities = iterateEntities(getPlayer().level(), createAABB(player.blockPosition().offset(calculateXLook(player) * 4.5, 1 + (calculateYLook(player) * 4.5), calculateZLook(player) * 4.5), 2));
-            entities.addAll(iterateEntities(getPlayer().level(), createAABB(player.blockPosition().offset(calculateXLook(player) * 3, 1 + (calculateYLook(player) * 3), calculateZLook(player) * 3), 2)));
-            entities.addAll(iterateEntities(getPlayer().level(), createAABB(player.blockPosition().offset(calculateXLook(player) * 1.5, 1 + (calculateYLook(player) * 1.5), calculateZLook(player) * 1.5), 2)));
+            List<Entity> entities = iterateEntities(getPlayer().level(), createAABB(player.blockPosition().offset((int) (calculateXLook(player) * 4.5), (int) (1 + (calculateYLook(player) * 4.5)), (int) (calculateZLook(player) * 4.5)), 2));
+            entities.addAll(iterateEntities(getPlayer().level(), createAABB(player.blockPosition().offset((int) (calculateXLook(player) * 3), (int) (1 + (calculateYLook(player) * 3)), (int) (calculateZLook(player) * 3)), 2)));
+            entities.addAll(iterateEntities(getPlayer().level(), createAABB(player.blockPosition().offset((int) (calculateXLook(player) * 1.5), (int) (1 + (calculateYLook(player) * 1.5)), (int) (calculateZLook(player) * 1.5)), 2)));
 
             if (entities.size() > 0) getPlayer().playSound(CSSoundEvents.CS_BLING.get(), 0.15F, 1F + getPlayer().level().random.nextFloat());
